% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delays.r
\name{add_delays}
\alias{add_delays}
\title{Simulate symptom onset and hospitalization times from infection times}
\usage{
add_delays(infection_times)
}
\arguments{
\item{infection_times}{A data frame containing a column of infection times}
}
\value{
A data frame with columns for infection time, onset time, and
  hospitalization time (with 70% of hospitalizations set to NA)
}
\description{
Simulate symptom onset and hospitalization times from infection times
}
\examples{
delayed_infections <- add_delays(infection_times)
head(delayed_infections)
}
