% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored-delay-pmf.r
\name{censored_delay_pmf}
\alias{censored_delay_pmf}
\title{Discretise a Continuous Delay Distribution}
\usage{
censored_delay_pmf(rgen, max, n = 1e+06, ...)
}
\arguments{
\item{rgen}{A function that generates random delays, e.g., `rgamma`,
`rlnorm`.}

\item{max}{The maximum delay.}

\item{n}{The number of replicates to simulate. Defaults to `1e+6`.}

\item{...}{Additional parameters of the delay distribution.}
}
\value{
A vector of probabilities corresponding to discrete indices from
  `0` to `max`, representing the discretised delay distribution.
}
\description{
This function discretises a continuous delay distribution into a probability
mass function (PMF) over discrete days.
}
\examples{
censored_delay_pmf(rgen = rgamma, max = 14, shape = 5, rate = 1)
}
