% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-onsets-by-report.r
\name{condition_onsets_by_report}
\alias{condition_onsets_by_report}
\title{Reduces the number of symptom onsets by a delay distribution.}
\usage{
condition_onsets_by_report(onsets, delay)
}
\arguments{
\item{onsets}{vector of symptom onsets}

\item{delay}{vector representing the cumulative distribution function (CDF)
of the (deiscrete) delay distribution. The first element corresponds to a
delay of 0, i.e. symptom onsets reported on the same day.}
}
\value{
A vector of symptom onsets reduced by the delay distribution.
}
\description{
This function assumes that symptom onsets are reported with a delay. It
reduces the number of symptom onsets by this delay, assuming that the last
element of the symptom onset vector is the day of reporting.
}
