% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve-with-delay.r
\name{convolve_with_delay}
\alias{convolve_with_delay}
\title{Convolve a Time Series with a Delay Distribution}
\usage{
convolve_with_delay(ts, delay_pmf)
}
\arguments{
\item{ts}{A vector of the time series to convolve.}

\item{delay_pmf}{The probability mass function of the delay, given as a
vector of probabilities, corresponding to discrete indices 0, 1, 2 of the
discretised delay distribution.}
}
\value{
A vector of the convolved time series.
}
\description{
This function convolves a time series with a delay distribution given as a
probability mass function (PMF).
}
\examples{
convolve_with_delay(ts = c(10, 14, 10, 10), delay_pmf = c(0.1, 0.6, 0.3))
}
