% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{covid_locations}
\alias{covid_locations}
\title{Data on forecasted locations from the US COVID-19 Forecast Hub}
\format{
A [tibble::tibble()] with a 4 columns and 53 rows
\describe{
  \item{abbreviation}{two-letter abbreviation for a jurisdiction}
  \item{location}{a two-letter FIPS code for identifying the jurisdiction}
  \item{location_name}{name of the jurisdiction}
  \item{population}{population of the jurisdiction as of some unknown time}
}
}
\usage{
covid_locations
}
\description{
A dataset containing 53 rows with metadata about each US state.
The dataset was downloaded from
https://raw.githubusercontent.com/CDCgov/covid19-forecast-hub/refs/heads/main/auxiliary-data/locations.csv
}
\keyword{datasets}
