% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{covid_time_series}
\alias{covid_time_series}
\title{Time-series data on COVID-19 in the US}
\format{
A [tibble::tibble()] with a 6 columns and 52,778 rows.
\describe{
  \item{date}{date of the Saturday ending an epidemic week for which the
  count was made}
  \item{state}{two-letter abbreviation for a jurisdiction}
  \item{observation}{the count of the observation, e.g. number of hospital
  admissions in that week and location}
  \item{location}{a two-letter FIPS code for identifying the jurisdiction}
  \item{as_of}{date corresponding to the date (typically a Wednesday) the
  data were archived}
  \item{target}{the name of the prediction target: "wk inc covid hosp"}
  \item{abbreviation}{two letter abbreviation for the location}
  \item{location_name}{full name of the location}
  \item{population}{population size of the location}
}
}
\usage{
covid_time_series
}
\description{
A dataset containing versions of time-series data on COVID-19 in the US,
including data on weekly counts of hospital admissions due to COVID-19.
The first week of data for which observations are available ended on
2022-10-01 and the last week ended on 2025-06-28. This dataset was downloaded
on 2025-07-09, with versions as early as 2024-11-20 (when the hub started
recording this dataset).
}
\keyword{datasets}
