% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{.extract_stan_functions}
\alias{.extract_stan_functions}
\title{Extract function names or content from Stan code}
\usage{
.extract_stan_functions(content, names_only = FALSE, functions = NULL)
}
\arguments{
\item{content}{Character vector containing Stan code}

\item{names_only}{Logical, if TRUE extract function names, otherwise
extract function content.}

\item{functions}{Optional, character vector of function names to extract
content for.}
}
\value{
Character vector of function names or content
}
\description{
Extract function names or content from Stan code
}
\keyword{internal}
