% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric-diff-ar.r
\name{geometric_diff_ar}
\alias{geometric_diff_ar}
\title{Geometric Differenced Autoregressive Process}
\usage{
geometric_diff_ar(init, noise, std, damp)
}
\arguments{
\item{init}{The initial value.}

\item{noise}{A vector of steps (on the standard normal scale).}

\item{std}{The step size of the random walk.}

\item{damp}{A damping parameter.}
}
\value{
A vector of the generated geometric differenced autoregressive
process.
}
\description{
This function generates a geometric differenced autoregressive process.
}
\examples{
geometric_diff_ar(init = 1, noise = rnorm(100), phi = 0.1, std = 0.1)
}
