% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric-random-walk.r
\name{geometric_random_walk}
\alias{geometric_random_walk}
\title{Geometric Random Walk}
\usage{
geometric_random_walk(init, noise, std)
}
\arguments{
\item{init}{The initial value.}

\item{noise}{A vector of steps (on the standard normal scale).}

\item{std}{The step size of the random walk.}
}
\value{
A vector of the generated geometric random walk.
}
\description{
This function generates a geometric random walk.
}
\examples{
geometric_random_walk(init = 1, noise = rnorm(100), std = 0.1)
}
