% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{infection_times}
\alias{infection_times}
\title{Infection times}
\format{
A data frame with a single column
\describe{
  \item{infection_time}{the times at which individuals were infected (and became infectious)}
}
}
\usage{
infection_times
}
\description{
A dataset containing random infection times from a branching process model,
generated using the code in data-raw/epicurve.r
}
\keyword{datasets}
