% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-daily-infections.r
\name{make_daily_infections}
\alias{make_daily_infections}
\title{Convert infection times to a daily time series}
\usage{
make_daily_infections(infection_times)
}
\arguments{
\item{infection_times}{A data frame containing a column of infection times}
}
\value{
A data frame with columns infection_day and infections, containing
  the daily count of infections
}
\description{
Convert infection times to a daily time series
}
\examples{
make_daily_infections(infection_times)
}
