% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_onsets.r
\name{make_gen_time_pmf}
\alias{make_gen_time_pmf}
\title{Generate a probability mass function for the generation time}
\usage{
make_gen_time_pmf(max = 14, shape = 4, rate = 1)
}
\arguments{
\item{max}{Maximum delay to consider}

\item{shape}{Shape parameter for the gamma distribution}

\item{rate}{Rate parameter for the gamma distribution}
}
\value{
A vector of probabilities representing the generation time PMF
}
\description{
Generate a probability mass function for the generation time
}
