% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mech_forecasts}
\alias{mech_forecasts}
\title{Forecasts from a mechanistic model}
\format{
A [tibble::tibble()] with a 7 columns
\describe{
  \item{day}{the day for which the forecast was made}
  \item{.draw}{an ID of the posterior predictive sample}
  \item{.variable}{name of the variable}
  \item{.value}{predicted value}
  \item{.horizon}{the forecast horizon in days}
  \item{target_day}{the day on which the forecast was made (using data up to this day)}
  \item{model}{the name of the model}
}
}
\usage{
mech_forecasts
}
\description{
A dataset containing forecasts from a mechanistic model, generated using the
code in data-raw/generate-example-forecasts.r
}
\keyword{datasets}
