% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_load_stan_functions}
\alias{nfidd_load_stan_functions}
\title{Load Stan functions as a string}
\usage{
nfidd_load_stan_functions(
  functions = NULL,
  stan_path = nfidd::nfidd_stan_path(),
  wrap_in_block = FALSE,
  write_to_file = FALSE,
  output_file = "nfidd_functions.stan"
)
}
\arguments{
\item{functions}{Character vector of function names to load. Defaults to all
functions.}

\item{stan_path}{Character string, the path to the Stan code. Defaults to the
path to the Stan code in the nfidd package.}

\item{wrap_in_block}{Logical, whether to wrap the functions in a
`functions{}` block. Default is FALSE.}

\item{write_to_file}{Logical, whether to write the output to a file. Default
is FALSE.}

\item{output_file}{Character string, the path to write the output file if
write_to_file is TRUE. Defaults to "nfidd_functions.stan".}
}
\value{
A character string containing the requested Stan functions
}
\description{
Load Stan functions as a string
}
\seealso{
Other stantools: 
\code{\link{nfidd_stan_function_files}()},
\code{\link{nfidd_stan_functions}()},
\code{\link{nfidd_stan_path}()}
}
\concept{stantools}
