% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_sample}
\alias{nfidd_sample}
\title{Sample from a CmdStanModel with NFIDD course defaults}
\usage{
nfidd_sample(
  model,
  iter_warmup = 500,
  iter_sampling = 500,
  parallel_chains = 4,
  ...
)
}
\arguments{
\item{model}{A CmdStanModel object to sample from.}

\item{iter_warmup}{Integer, number of warmup iterations per chain.
Defaults to 500 (reduced from cmdstanr default of 1000) for course speed.}

\item{iter_sampling}{Integer, number of sampling iterations per chain.
Defaults to 500 (reduced from cmdstanr default of 1000) for course speed.}

\item{parallel_chains}{Integer, number of chains to run in parallel.
Defaults to 4 for course speed.}

\item{...}{Additional arguments passed to the model's sample method.
All cmdstanr sample arguments are supported.}
}
\value{
A CmdStanMCMC object containing the posterior samples.
}
\description{
This function wraps the cmdstanr sample method with optimized defaults
for course use to speed up model fitting. All cmdstanr sample arguments
can still be overridden for experimentation.
}
\seealso{
Other modelhelpers: 
\code{\link{nfidd_cmdstan_model}()}
}
\concept{modelhelpers}
