% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_stan_functions}
\alias{nfidd_stan_functions}
\title{Get Stan function names from Stan files}
\usage{
nfidd_stan_functions(stan_path = nfidd::nfidd_stan_path())
}
\arguments{
\item{stan_path}{Character string specifying the path to the directory
containing Stan files. Defaults to the Stan path of the nfidd
package.}
}
\value{
A character vector containing unique names of all functions found in
the Stan files.
}
\description{
This function reads all Stan files in the specified directory and extracts
the names of all functions defined in those files.
}
\seealso{
Other stantools: 
\code{\link{nfidd_load_stan_functions}()},
\code{\link{nfidd_stan_function_files}()},
\code{\link{nfidd_stan_path}()}
}
\concept{stantools}
