% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_stan_models}
\alias{nfidd_stan_models}
\title{List Available Stan Models in NFIDD}
\usage{
nfidd_stan_models(stan_path = nfidd::nfidd_stan_path())
}
\arguments{
\item{stan_path}{Character string specifying the path to Stan files. Defaults
to the result of `nfidd_stan_path()`.}
}
\value{
A character vector of available Stan model names.
}
\description{
This function finds all available Stan models in the NFIDD package and
returns their names without the .stan extension.
}
\examples{
nfidd_list_stan_models()
}
