% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_stan_path}
\alias{nfidd_stan_path}
\title{Get the path to Stan code}
\usage{
nfidd_stan_path()
}
\value{
A character string with the path to the Stan code
}
\description{
Get the path to Stan code
}
\seealso{
Other stantools: 
\code{\link{nfidd_load_stan_functions}()},
\code{\link{nfidd_stan_function_files}()},
\code{\link{nfidd_stan_functions}()}
}
\concept{stantools}
