% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal.r
\name{renewal}
\alias{renewal}
\title{Simulate Infections using the Renewal Equation}
\usage{
renewal(I0, R, gen_time)
}
\arguments{
\item{I0}{The initial number of infections.}

\item{R}{The reproduction number, given as a vector with one entry per time
point.}

\item{gen_time}{The generation time distribution, given as a vector with one
entry per day after infection (the first element corresponding to one day
after infection).}
}
\value{
A vector of simulated infections over time.
}
\description{
This function simulates infections using the renewal equation.
}
\examples{
renewal(
  I0 = 5,
  R = c(rep(3, 4), rep(0.5, 5)),
  gen_time = c(0.1, 0.2, 0.3, 0.2, 0.1)
)
}
