% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rw_forecasts}
\alias{rw_forecasts}
\title{Forecasts from a semi-mechanistic model}
\format{
A [tibble::tibble()] with a 7 columns
\describe{
  \item{day}{the day for which the forecast was made}
  \item{.draw}{an ID of the posterior predictive sample}
  \item{.variable}{name of the variable}
  \item{.value}{predicted value}
  \item{.horizon}{the forecast horizon in days}
  \item{target_day}{the day on which the forecast was made (using data up to this day)}
  \item{model}{the name of the model}
}
}
\usage{
rw_forecasts
}
\description{
A dataset containing forecasts from a semi-mechanistic model (using a
geometric random walk prior on the reproduction number), generated using the
code in data-raw/generate-example-forecasts.r
}
\keyword{datasets}
