% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{stat_forecasts}
\alias{stat_forecasts}
\title{Forecasts from a semi-mechanistic model with additional statistical
complexity}
\format{
A [tibble::tibble()] with a 7 columns
\describe{
  \item{day}{the day for which the forecast was made}
  \item{.draw}{an ID of the posterior predictive sample}
  \item{.variable}{name of the variable}
  \item{.value}{predicted value}
  \item{.horizon}{the forecast horizon in days}
  \item{origin_day}{the day on which the forecast was made (using data up to
  this day)}
  \item{model}{the name of the model}
}
}
\usage{
stat_forecasts
}
\description{
A dataset containing forecasts from a semi-mechanistic model (using an
autoregressive prior for reproduction number), generated using the code in
data-raw/generate-example-forecasts.r
}
\keyword{datasets}
