% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{covid_forecasts}
\alias{covid_forecasts}
\title{Forecast data for COVID-19 hospitalizations in the US}
\format{
A [tibble::tibble()] with a 9 columns and 2,002,633 rows.
\describe{
  \item{reference_date}{the date of the Saturday following the Wednesday
  forecast submission date each week}
  \item{location}{a two-letter FIPS code for identifying the jurisdiction}
  \item{horizon}{an integer between 0 and 3, corresponding to the
  prediction horizon}
  \item{target_end_date}{date corresponding to the Saturday ending the
  epiweek being forecasted}
  \item{target}{"wk inc covid hosp", the name of the forecast target,
  corresponding to weekly incident covid hospitalizations}
  \item{output_type}{string denoting the forecast format, here, all "quantile"}
  \item{output_type_id}{number between 0 and 1 indicating the quantile level
  corresponding to the prediction}
  \item{value}{the numeric value of the prediction at the specified quantile
  level}
  \item{model_id}{the unique character string identifying the model that
  made the prediction}
  \item{abbreviation}{two letter abbreviation for the location}
  \item{location_name}{full name of the location}
  \item{population}{population size of the location}
}
}
\usage{
covid_forecasts
}
\description{
A dataset containing forecasts of hospital admission counts in the US from
November 2024 through early July 2025.
}
\keyword{datasets}
