% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_cmdstan_model}
\alias{nfidd_cmdstan_model}
\title{Create a CmdStanModel with NFIDD Stan functions}
\usage{
nfidd_cmdstan_model(model_name, include_paths = nfidd::nfidd_stan_path(), ...)
}
\arguments{
\item{model_name}{Character string specifying which Stan model to use.}

\item{include_paths}{Character vector of paths to include for Stan
compilation. Defaults to the result of `nfidd_stan_path()`.}

\item{...}{Additional arguments passed to cmdstanr::cmdstan_model().}
}
\value{
A CmdStanModel object.
}
\description{
This function creates a CmdStanModel object using a specified Stan model from
the NFIDD package and optionally includes additional user-specified Stan
files.
}
\examples{
\dontshow{if (requireNamespace("cmdstanr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
if (!is.null(cmdstanr::cmdstan_version(error_on_NA = FALSE))) {
  model <- nfidd_cmdstan_model("renewal", compile = FALSE)
  model
}
\dontshow{\}) # examplesIf}
}
\concept{modelhelpers}
