% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-lognormal.r
\name{summarise_lognormal}
\alias{summarise_lognormal}
\title{Summarise lognormal distribution parameters}
\usage{
summarise_lognormal(posterior_obj)
}
\arguments{
\item{posterior_obj}{A posterior or draws object containing lognormal
parameters}
}
\value{
A summary of the mean and standard deviation of the lognormal
  distribution
}
\description{
Extract mean and standard deviation from lognormal distribution parameters
in a posterior or draws object.
}
\details{
This function extracts lognormal parameters from a posterior object (expcting
  them to be called `meanlog` and `sdlog`and calculates the mean and standard
  deviation on the natural scale using:
\itemize{
\item \code{mean = exp(meanlog + 0.5 * sdlog^2)}
\item \code{sd = exp(meanlog + 0.5 * sdlog^2) * sqrt(exp(sdlog^2) - 1)}
}
}
\examples{
\dontrun{
# After fitting a lognormal model with cmdstan
res |> summarise_lognormal()
}
}
